feat: 참가자 공통 기능 추상화
